Use https://golang.org/dl/?mode=json as input
Retrieve the list of current versions from
https://golang.org/dl/?mode=json and select the filename and sha256
value from there, depending on the version requested.

The calculated hash of the downloaded file is checked against the value
in the JSON-blob.

Commandline parameters now also include -arch, -os and -n (for dry run).
The final string in the command line is optional and might be the
version we are interested in.

Fixes #1