Get version of local go installation
Previous to this commit, goup took the output of `runtime.Version()` as
reference.  However, that value wouldn't change as it it bound to the
runtime of the running binary of `goup`.

With this commit we parse the output of `go version` to determine the
version of the local go installation.